requirements.txt
streamlit>=1.40.0
torch>=2.4.0
torchvision>=0.19.0
Pillow>=10.0.0
numpy>=1.26.0
opencv-python-headless>=4.10.0
matplotlib>=3.8.0
pandas>=2.2.0
